requirements.txt
Flask
waitress
python-dotenv
oracledb
Werkzeug